Added a better Makefile for the McGill Linux server

+++ src/iterative-alpha-beta.h
@@ -194,6 +194,6 @@ private:
         auto timeInMs = std::chrono::duration_cast<std::chrono::milliseconds>(
                             now - startTime)
                             .count();
-        return timeInMs >= timeLimitInMs;
+        return static_cast<size_t>(timeInMs) >= timeLimitInMs;
     }
 };
